+++ src/jobs/SendoutJob.php
@@ -124,12 +124,12 @@ class SendoutJob extends BaseJob implements RetryableJobInterface
         $pendingRecipients = $sendout->getPendingRecipients();
 
         $count = 0;
+        $expectedRecipients = count($pendingRecipients);
 
         // Loop as long as the there are pending recipients and the sendout is sendable
         while (count($pendingRecipients) AND $sendout->isSendable()) {
             // Set progress
-            $progress = $sendout->getProgressFraction();
-            $this->setProgress($queue, $progress);
+            $this->setProgress($queue, $count / $expectedRecipients);
 
             // Get next pending recipient
             $pendingRecipient = array_shift($pendingRecipients);