+++ src/records/SendoutRecord.php
@@ -25,9 +25,8 @@ use DateTime;
  * @property boolean     $googleAnalyticsLinkTracking Google Analytics link tracking
  * @property string      $mailingListIds             Mailing list IDs
  * @property string      $excludedMailingListIds     Excluded mailing list IDs
- * @property int         $expectedRecipients         Expected recipients
- * @property int         $failedRecipients           Failed recipients
  * @property int         $recipients                 Recipients
+ * @property int         $failedRecipients           Failed recipients
  * @property mixed       $automatedSchedule          Automated schedule
  * @property DateTime    $sendDate                   Send date
  * @property DateTime    $lastSent                   Last sent